want to implemnt DDC in software for 100ms data. then compare channel A and B amplitudes.
I'm not sure what to do with the integration bandwidths for the jitter calculation. I can repoduce the analog paper's numebers and my numbers are ball park reasonable, but still not correct. I think the part that is confusing is the 1kHz integration bandwidth, due to the helicity flip time. This spectral effect is already included in the FFT (i think), because the FFT number of points is limited to the 1ms window. I don't know if this should be further accounted for in the integration limits for the jitter due to the system's phase noise.